refactor(archtest): Outbox/Relay/Projection 规则族迁 module-path-agnostic + 缩 baseline [#1635]
M3 PR-5（EPIC #1302）：9 个 Outbox/Relay/Projection 族 archtest 规则文件的平台路径
硬编码经 PlatformModulePath 派生，离开 module_path_funnel.baseline（64→55）。

形态跟随功能：
- 外部适用纯禁令规则 → L3 Check* + 注册 StandardCellRules（外部仓库可跑）：
  OUTBOX-RECONSTRUCTION-CALLER-01 / PROJECTION-APPLY-HOOK-FUNNEL-01 /
  OUTBOX-HANDLERESULT-FACTORY-PREFERRED-01。
- gocell self-check 规则 → 就地把平台路径字面量派生自 PlatformModulePath，逻辑留
  _test.go（不造 .go/Check* 平行结构）：publisher-enrollment / outboxtest-import /
  l2-atomicity / relay-isolation / checkpoint-conformance-enroll / checkpoint-tx-bound /
  OUTBOXTEST-CLOSE-VIA-BUDGET-01。
- OUTBOX-TOPIC-FAILOPEN-01 → 迁 .go Check* 但不注册：kernel/outbox.Entry sealed，
  生产无法构造其字面量，扫描在 gocell 与外部均 vacuous，仅 fixture 覆盖（同
  SCAFFOLD-LISTENER-MARKER 先例）。
- 11 个无平台字面量的 outbox 规则零改动（已 module-path-agnostic）。

行为零变：既有 dogfood + RED fixture 全绿；TestArchtestModulePathFunnel ratchet 通过。

ref: golang.org/x/tools/go/analysis (CellRule = Analyzer 形态)
