+++ tools/archtest/l2_outbox_atomicity_coverage_test.go
@@ -92,7 +92,7 @@ import (
 const ruleL2AtomicityCoverage = "L2-OUTBOX-ATOMICITY-COVERAGE-01"
 
 // outboxPkgImportPath is the canonical import path of kernel/outbox.
-const outboxPkgImportPath = "github.com/ghbvf/gocell/kernel/outbox"
+const outboxPkgImportPath = PlatformModulePath + "/kernel/outbox"
 
 // l2MinUnits is the floor assertion: scan must find at least this many L2
 // units. Fail-closed against a broken YAML scan silently zeroing coverage.
