+++ tools/archtest/testdata/module_path_funnel.baseline
@@ -33,28 +33,19 @@ tools/archtest/health_aggregation_test.go
 tools/archtest/health_verbose_invariants_test.go
 tools/archtest/healthz_invariants_test.go
 tools/archtest/kernel_poolstats_location_test.go
-tools/archtest/l2_outbox_atomicity_coverage_test.go
 tools/archtest/metrics_cancel_ctx_conformance_test.go
 tools/archtest/mqtt_funnel_test.go
 tools/archtest/mqtt_reason_redaction_test.go
 tools/archtest/no_deleted_auth_symbols_test.go
 tools/archtest/notfound_test_strict_test.go
 tools/archtest/observability_metrics_test.go
-tools/archtest/outbox_invariants_test.go
-tools/archtest/outbox_publisher_conformance_enrollment_test.go
-tools/archtest/outbox_reconstruction_caller_test.go
-tools/archtest/outboxtest_import_boundary_test.go
 tools/archtest/parser_matcher_examples_symmetry_test.go
 tools/archtest/pass_funnel_test.go
 tools/archtest/pg_repo_ambient_tx_test.go
 tools/archtest/probename_sealed_funnel_test.go
-tools/archtest/projection_apply_hook_funnel_test.go
-tools/archtest/projection_checkpoint_conformance_enroll_test.go
-tools/archtest/projection_checkpoint_tx_bound_test.go
 tools/archtest/prom_cell_label_funnel_test.go
 tools/archtest/reconstitute_user_caller_allowlist_test.go
 tools/archtest/refresh_invariants_test.go
-tools/archtest/relay_isolation_test.go
 tools/archtest/required_dep_nil_guard_test.go
 tools/archtest/reverse_coverage_invariants_test.go
 tools/archtest/rmq_invariants_test.go
